requirements.txt
streamlit
google-genai
pandas
st-gsheets-connection



